+++ openmls/src/group/tests/test_commit_validation.rs
@@ -5,7 +5,7 @@ use openmls_rust_crypto::OpenMlsRustCrypto;
 use openmls_traits::{signatures::Signer, types::Ciphersuite};
 use rstest::*;
 use rstest_reuse::{self, *};
-use tls_codec::Serialize;
+use tls_codec::{Deserialize, Serialize};
 
 use super::utils::{
     generate_credential_with_key, generate_key_package, resign_message, CredentialWithKeyAndSigner,